Change subdirectories of Middleware
