+++ app/Http/Middleware/Game/CheckPeriod.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Middleware;
+namespace App\Http\Middleware\Game;
 
 use App\Models\Game\Game;
 use Closure;
