+++ app/Http/Middleware/User/RedirectToRoleDashboard.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Middleware;
+namespace App\Http\Middleware\User;
 
 use App\Models\User\User;
 use Closure;
